3/16 Update for Cult of Pleasure
Update that adds a unique assassination mission system for the Cult of Pleasure as well as a power for the Grandmaster to mark someone for death. Sets Morathi to be the starting Grandmaster. Makes several loc fixes for cult of pleasure related events. Turns the Cult of Pleasure into a Dual Society that changes into a secret Slaanesh cult for Cadai elves of rank 3 or greater and all elves of rank 3 or greater if Morathi is killed.
Gives access to certain secret cult missions that make sense.

There is also a fix for the chaos cults being able to create secret communities in counties that openly worship the chaos god already, as long as that god wasn't Khorne.